fix: address PR #3460 review feedback and CI failure
- Renumber ADR-0015 → ADR-0018 in heading, index.md href, and mkdocs nav
  (fixes test_adr_heading_matches_filename CI failure after the file rename).
- WhatsAppMessage.determine_content_type: add `case _` fallback returning
  MESSAGE_TYPES.OTHER so unknown provider types never yield None.
- Move WhatsappAttachmentHydrationStage from stages/core.py into
  whatsapp_channel.py (co-located with its channel); update test imports.
- Guard nullable messaging_provider in the hydration stage's _get_files to
  avoid AttributeError when a WhatsApp channel has no provider.
- Extract duplicated get_inbound_media/get_message_audio from Twilio, Turn.io,
  and Meta services into HttpMediaDownloadMixin; the shared handler now catches
  both requests.HTTPError and httpx.HTTPStatusError, removing prior
  inconsistency.
