+++ apps/channels/channels_v2/stages/core.py
@@ -699,71 +699,3 @@ class AttachmentHydrationStage(ProcessingStage):
                 team_id=ctx.experiment.team_id,
             )
         )
-
-
-# ---------------------------------------------------------------------------
-# WhatsappAttachmentHydrationStage
-# ---------------------------------------------------------------------------
-
-
-class WhatsappAttachmentHydrationStage(AttachmentHydrationStage):
-    """Download, persist, and hydrate inbound WhatsApp media attachments
-    (images, documents, and other non-voice media).
-
-    should_run fires only when the message references downloadable media —
-    WhatsAppMessage.parse sets attachment_mime_type on every parsed message
-    (including text), so the mime type alone isn't a reliable gate.
-    _get_files() downloads the media via the messaging service and persists
-    the bytes as a MESSAGE_MEDIA File. The base class then handles
-    ChatAttachment linkage and Attachment construction. Size and
-    content-type policing is the upstream provider's responsibility —
-    Meta already caps what reaches us.
-    """
-
-    def should_run(self, ctx: MessageProcessingContext) -> bool:
-        if not ctx.experiment_session or not ctx.message:
-            return False
-        if ctx.message.attachments:
-            return False
-        if not ctx.message.attachment_mime_type:
-            return False
-        return bool(getattr(ctx.message, "media_url", None) or getattr(ctx.message, "media_id", None))
-
-    def _get_files(self, ctx: MessageProcessingContext) -> list[File]:
-        messaging_service = ctx.experiment_channel.messaging_provider.get_messaging_service()
-
-        try:
-            media = messaging_service.get_inbound_media(ctx.message)
-        except Exception:
-            logger.exception("Failed to download WhatsApp inbound media")
-            return []
-
-        if media is None:
-            return []
-
-        raw_bytes, content_type = media
-        filename = self._resolve_filename(ctx, content_type)
-
-        try:
-            file = File.create(
-                filename=filename,
-                file_obj=BytesIO(raw_bytes),
-                team_id=ctx.experiment.team_id,
-                purpose=FilePurpose.MESSAGE_MEDIA,
-                content_type=content_type,
-            )
-        except Exception:
-            logger.exception("Failed to persist WhatsApp inbound media")
-            return []
-
-        return [file]
-
-    @staticmethod
-    def _resolve_filename(ctx: MessageProcessingContext, content_type: str) -> str:
-        """Prefer the provider-supplied filename (documents). Fall back to a
-        family-based name when the provider doesn't send one (e.g. images)."""
-        provided = getattr(ctx.message, "attachment_filename", None)
-        if provided:
-            return provided
-        family = content_type.split("/", 1)[0] if "/" in content_type else "attachment"
-        return family or "attachment"
