+++ docs/adr/0018-download-whatsapp-inbound-attachments-in-hydration-stage.md
@@ -1,4 +1,4 @@
-# ADR-0015: Download WhatsApp inbound attachments in an overridden hydration stage
+# ADR-0018: Download WhatsApp inbound attachments in an overridden hydration stage
 
 <span class="adr-status adr-status-accepted">ACCEPTED</span>
